+++ packages/expo-ui/ios/ExpoUIModule.swift
@@ -5,7 +5,6 @@ import ExpoModulesCore
 public class ExpoUIModule: Module {
   public func definition() -> ModuleDefinition {
     Name("ExpoUI")
-
     View(Button.self)
     View(PickerView.self)
     View(SwitchView.self)
@@ -17,5 +16,7 @@ public class ExpoUIModule: Module {
     View(TextInputView.self)
     View(ProgressView.self)
     View(GaugeView.self)
+    View(ListView.self)
+    View(LabelView.self)
   }
 }